+++ snappylapy/expectation_classes/__init__.py
@@ -4,6 +4,7 @@ from .expect_bytes import BytesExpect
 from .expect_dataframe import DataframeExpect
 from .expect_dict import DictExpect
 from .expect_list import ListExpect
+from .expect_object import ObjectExpect
 from .expect_string import StringExpect
 
 __all__ = [
@@ -11,5 +12,6 @@ __all__ = [
     "DataframeExpect",
     "DictExpect",
     "ListExpect",
+    "ObjectExpect",
     "StringExpect",
 ]