+++ snappylapy/expectation_classes/base_snapshot.py
@@ -39,7 +39,7 @@ class BaseSnapshot(ABC, Generic[T]):
         """Assert test results match the snapshot."""
         if not (self.settings.snapshot_dir / self.settings.filename).exists():
             if not self.settings.snapshot_update:
-                error_msg = f"Snapshot file not found: {self.settings.filename}, run pytest with the --snapshot-update flag to create it."  # noqa: E501
+                error_msg = f"Snapshot file not found: {self.settings.filename}, run 'snappylapy update' command in the terminal, or run pytest with the --snapshot-update flag to create it."  # noqa: E501
                 raise FileNotFoundError(error_msg)
             self.snappylapy_session.add_created_snapshot(
                 self.settings.filename)
@@ -52,7 +52,9 @@ class BaseSnapshot(ABC, Generic[T]):
                                     self.settings.filename)
         try:
             snapshot_data_str = snapshot_data.decode()
+            snapshot_data_str = snapshot_data_str.replace("\r\n", "\n").replace("\r", "\n")
             test_data_str = test_data.decode()
+            test_data_str = test_data_str.replace("\r\n", "\n").replace("\r", "\n")
             assert snapshot_data_str == test_data_str
         except AssertionError as error:
             if self.settings.snapshot_update: